# swagger/swagger.yaml
definitions:
  github_com_wesleyburlani_go-rest-api_internal_users.User:
    properties:
      email:
        type: string
      id:
        type: integer
      password:
        type: string
    type: object
  github_com_wesleyburlani_go-rest-api_pkg_crypto.JwtToken:
    properties:
      token:
        type: string
    type: object
  users.PostBody:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.PutBody:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
  /auth/login:
    post:
      description: generates a jwt token for a user
      parameters:
      - description: login user properties
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.PostBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_wesleyburlani_go-rest-api_pkg_crypto.JwtToken'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: generates a jwt token for a user
      tags:
      - auth
  /users:
    get:
      description: lists existing users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_wesleyburlani_go-rest-api_internal_users.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: lists existing users
      tags:
      - users
    post:
      description: creates a new user
      parameters:
      - description: create user properties
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.PostBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_wesleyburlani_go-rest-api_internal_users.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: creates a new user
      tags:
      - users
  /users/{id}:
    get:
      description: returns an existing user
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_wesleyburlani_go-rest-api_internal_users.User'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: returns an existing user
      tags:
      - users
    put:
      description: updates an existing user
      parameters:
      - description: user id
        in: path
        name: id
        required: true
        type: integer
      - description: update user properties
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/users.PutBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_wesleyburlani_go-rest-api_internal_users.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: updates an existing user
      tags:
      - users
  /users/me:
    get:
      description: returns the current authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_wesleyburlani_go-rest-api_internal_users.User'
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: returns the current authenticated user
      tags:
      - users
swagger: "2.0"